style: 폴더 구조 변경, AccountList Motion 추가
